fixed small testing bug in TraceFitter tests and conversion warning in TraceFitter base
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@6927 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/TRANSFORMATIONS/FEATUREFINDER/TraceFitter.h
@@ -255,7 +255,7 @@ namespace OpenMS
       T = gsl_multifit_fdfsolver_lmsder;
       s = gsl_multifit_fdfsolver_alloc(T, data_count, num_params);
       gsl_multifit_fdfsolver_set(s, &func, &x.vector);
-      size_t iter = 0;
+      Int iter = 0;
       do
       {
         iter++;
